fix bug

+++ trep/bl.c
@@ -171,7 +171,7 @@ void sum(unit *uptr) {
 	char result_str[64] = "";
 
 	for (int i = 0; i < uptr->child_num; i++)
-		result += atoi(get_child(uptr, i)->value);
+		result += atof(get_child(uptr, i)->value);
 
 	sprintf(result_str, "%d", result);
 	strcpy(uptr->value, result_str);
@@ -182,10 +182,10 @@ void sub(unit *uptr) {
 	char result_str[64] = "";
 
 	if (uptr->child_num > 1)
-		result = atoi(get_child(uptr, 0)->value);
+		result = atof(get_child(uptr, 0)->value);
 
 	for (int i = 1; i < uptr->child_num; i++)
-		result -= atoi(get_child(uptr, i)->value);
+		result -= atof(get_child(uptr, i)->value);
 
 	sprintf(result_str, "%d", result);
 	strcpy(uptr->value, result_str);
@@ -196,7 +196,7 @@ void mul(unit *uptr) {
 	char result_str[64] = "";
 
 	for (int i = 0; i < uptr->child_num; i++)
-		result *= atoi(get_child(uptr, i)->value);
+		result *= atof(get_child(uptr, i)->value);
 
 	sprintf(result_str, "%d", result);
 	strcpy(uptr->value, result_str);
@@ -207,10 +207,10 @@ void divop(unit *uptr) {
 	char result_str[64] = "";
 
 	if (uptr->child_num > 1)
-		result = atoi(get_child(uptr, 0)->value);
+		result = atof(get_child(uptr, 0)->value);
 
 	for (int i = 1; i < uptr->child_num; i++)
-		result /= atoi(get_child(uptr, i)->value);
+		result /= atof(get_child(uptr, i)->value);
 
 	sprintf(result_str, "%d", result);
 	strcpy(uptr->value, result_str);
